requirements.txt
trello==0.9.1
repoze.lru==0.6
